requirements.txt
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.1
django-allauth==0.43.0
django-bootstrap4==2.3.1
django-debug-toolbar==3.1.1
django-environ==0.4.5
idna==2.8
mysqlclient==2.0.1
oauthlib==3.1.0
pylint==1.8.1
pylint-django==0.8.0
pylint-plugin-utils==0.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
regex==2020.2.20
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
virtualenv==16.2.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
django-sendgrid-v5==0.9.0
